[IOTDB-3436] add aggregation with value filter  IT (#6389)
